gmail-listener/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
schedule
requests
